OPEN - issue BATCH-1401: All inserts of JobId should be of Types.BIGINT Applied patch to trunk

+++ spring-batch-core/src/main/java/org/springframework/batch/core/repository/dao/JdbcJobInstanceDao.java
@@ -117,7 +117,7 @@ public class JdbcJobInstanceDao extends AbstractJdbcBatchMetadataDao implements
 		getJdbcTemplate().getJdbcOperations().update(
 				getQuery(CREATE_JOB_INSTANCE),
 				parameters,
-				new int[] { Types.INTEGER, Types.VARCHAR, Types.VARCHAR,
+				new int[] { Types.BIGINT, Types.VARCHAR, Types.VARCHAR,
 						Types.INTEGER });
 
 		insertJobParameters(jobId, jobParameters);
